Skip deleted comments when parsing tickets.
If any comments are deleted on an issue, they'd still be included in the
resulting dump. If these issues contained attachments, the parser would
break.

We're now a bit more careful about how we're parsing the issue
description, comments, and file attachments. The parser will dive into
the proper elements and fetch only the data that's visible.